% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ones.R
\name{m.ones}
\alias{m.ones}
\title{Fills an array with ones as in Matlab}
\usage{
m.ones(x, ...)
}
\arguments{
\item{x}{1-D dimension}

\item{...}{additional parameters}
}
\description{
Fills an array with ones as in Matlab
}
