% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdiags.R
\name{m.spdiags}
\alias{m.spdiags}
\title{Extract and create sparse band and diagonal matrices (Octave)}
\usage{
m.spdiags(...)
}
\arguments{
\item{...}{additional parameters}
}
\description{
Extract and create sparse band and diagonal matrices (Octave)
}
