% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeros.R
\name{m.zeros}
\alias{m.zeros}
\title{Fills an array with zeros (Octave)}
\usage{
m.zeros(x, ...)
}
\arguments{
\item{x}{1-D dimension}

\item{...}{additional parameters}
}
\description{
Fills an array with zeros (Octave)
}
