% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndims.R
\name{ndims}
\alias{ndims}
\title{Number of dimensions of the object. As in Matlab}
\usage{
ndims(X)
}
\arguments{
\item{X}{vector, matrix or array}
}
\description{
Number of dimensions of the object. As in Matlab
}
