% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth3.R
\name{padreplicate}
\alias{padreplicate}
\title{Duplicate rows and columns in N-dimensional array}
\usage{
padreplicate(a, padSize)
}
\arguments{
\item{a}{n-dimensional array}

\item{padSize}{padding vector}
}
\description{
Duplicate rows and columns in N-dimensional array
}
