% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randn.R
\name{randn}
\alias{randn}
\title{random normal array}
\usage{
randn(..., mean = 0, sd = 1)
}
\arguments{
\item{...}{additional parameters. size of the array}

\item{mean}{as in rnorm}

\item{sd}{as in rnorm}
}
\description{
random normal array
}
