% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resave.R
\name{resave}
\alias{resave}
\title{Appends, saves variables to RDA file}
\usage{
resave(..., list = character(), file)
}
\arguments{
\item{...}{additional parameters}

\item{list}{a list}

\item{file}{a RDA file}
}
\description{
Appends, saves variables to RDA file
}
