% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreshape.R
\name{rreshape}
\alias{rreshape}
\title{Reshapes an array to a new n-dimensional array. Like Matlab}
\usage{
rreshape(x, ...)
}
\arguments{
\item{x}{array}

\item{...}{additional parameters}
}
\description{
Reshapes an array to a new n-dimensional array. Like Matlab
}
