% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth3.R
\name{Smooth3}
\alias{Smooth3}
\title{Smooth a 3-D array}
\usage{
Smooth3(data, filt = "box", sz = 3)
}
\arguments{
\item{data}{three dimensaional array}

\item{filt}{type of filter}

\item{sz}{size}
}
\description{
Smooth a 3-D array
}
