% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Grid.R
\name{Var3d}
\alias{Var3d}
\title{3D variable with x, y, z dimensions}
\usage{
Var3d(x = 0, y = 0, z = 0)
}
\arguments{
\item{x}{the x dimension}

\item{y}{the y dimension}

\item{z}{the z dimension}
}
\description{
3D variable with x, y, z dimensions
}
